Added GNU Windows Regex library
Adjusted some code to be able to use it without needing to include the
dll separately at compile time. Maybe this is wrong but it simplifies
work when just including the regex.c file.

Some adjustments are still needed since the compiler throws warnings of
redeclaration of functions. Also this needs to be tested.